add better indicator for no data on liveness page
