+++ packages/frontend/src/pages/scaling/liveness/types.ts
@@ -1,7 +1,5 @@
 import { Layer2, StageConfig } from '@l2beat/config'
 
-import { AnomalyIndicatorEntry } from '../../../components/AnomalyIndicator'
-
 interface DataPoint {
   averageInSeconds: number
   maximumInSeconds: number
@@ -11,6 +9,7 @@ export interface ScalingLivenessViewEntry {
   name: string
   slug: string
   category: Layer2['display']['category']
+  dataAvailabilityMode: Layer2['display']['dataAvailabilityMode']
   provider: Layer2['display']['provider'] | undefined
   warning: string | undefined
   stage: StageConfig
@@ -29,4 +28,19 @@ export interface ScalingLivenessViewEntry {
       }
     | undefined
   anomalyEntries: AnomalyIndicatorEntry[]
+}
+
+export type AnomalyIndicatorEntry = AnomalyEntry | NonAnomalyEntry
+export interface AnomalyEntry {
+  isAnomaly: true
+  anomalies: Anomaly[]
+}
+
+export interface NonAnomalyEntry {
+  isAnomaly: false
+}
+export interface Anomaly {
+  type: 'BATCH SUBMISSION' | 'STATE UPDATE'
+  timestamp: number
+  durationInSeconds: number
 }
